+++ Cargo.toml
@@ -7,6 +7,7 @@ edition = "2018"
 rand = "0.8.4"
 rayon = "1.5.1"
 lazy_static = "1.4.0"
+primes = "0.3.0"
 
 [dev-dependencies]
 criterion = "0.3"
