+++ benches/digraph_benchmark.rs
@@ -1,17 +1,19 @@
 use criterion::{black_box, criterion_group, criterion_main, Criterion};
 use graph::digraph::*;
 use graph::global::*;
+use graph::edge::*;
 use graph::examples::*;
 use graph::global::Traverse::{Include, Finish};
 use rand::Rng;
 use lazy_static::lazy_static;
+use std::sync::Arc;
 
-const SIMPLE_NODE_COUNT: usize = 1000000;
+const SIMPLE_NODE_COUNT: usize = 100000;
 const SIMPLE_NODE_DEGREE: usize = 10;
 const FLOW_NODE_COUNT: usize = 1000;
-const FLOW_NODE_DEGREE: usize = 10;
+const FLOW_NODE_DEGREE: usize = 20;
 
-type IntKeysGraph = Digraph<usize, Null, Null>;
+type IntKeysGraph = Digraph<usize, usize, Null>;
 
 fn create_graph_flow() -> FlowGraph {
 	let mut g = FlowGraph::new();
@@ -29,7 +31,7 @@ fn create_graph_flow() -> FlowGraph {
 fn create_graph_simple() -> IntKeysGraph {
 	let mut g = IntKeysGraph::new();
 	for i in 0..SIMPLE_NODE_COUNT {
-		g.insert(i, Null);
+		g.insert(i, rand_range(0, 10000000000));
 	}
 	for i in 0..SIMPLE_NODE_COUNT {
 		for _ in 0..SIMPLE_NODE_DEGREE {
@@ -52,7 +54,7 @@ fn rand_range(start: usize, end: usize) -> usize {
 fn create_graph_speed() {
 	let mut g = IntKeysGraph::new();
 	for i in 0..1000 {
-		g.insert(i, Null);
+		g.insert(i, rand_range(0, 10000000000));
 	}
 	for i in 0..1000 {
 		for _ in 0..100 {
@@ -69,8 +71,21 @@ fn digraph_breadth_first_search(c: &mut Criterion) {
 	// println!("constructing graph of size = {} Mb", ((SIMPLE_NODE_COUNT * std::mem::size_of::<Node<usize, usize, usize>>()) + (SIMPLE_NODE_COUNT * SIMPLE_NODE_DEGREE * std::mem::size_of::<Edge<usize, usize, usize>>())) / 1000_000);
 	fn digraph_bfs() {
 		let t = SIMPLE_GRAPH.node(&rand_range(0, SIMPLE_NODE_COUNT)).unwrap();
-		SIMPLE_GRAPH.breadth_first(&rand_range(0, SIMPLE_NODE_COUNT),
-	|e| if *t == e.target() { Finish } else { Include });
+		let closure = | e: &Arc<Edge<usize, usize, Null>> | {
+
+			let mut n = e.target().load();
+
+			if primes::is_prime(n as u64) == true {
+				n = n / 2;
+				e.target().store(n);
+			}
+			if *t == e.target() {
+				Finish
+			} else {
+				Include
+			}
+		};
+		SIMPLE_GRAPH.breadth_first(&rand_range(0, SIMPLE_NODE_COUNT), closure);
 	}
 	println!("graph node count = {}", SIMPLE_GRAPH.node_count());
 	println!("graph edge count = {}", SIMPLE_GRAPH.edge_count());
@@ -83,8 +98,21 @@ fn digraph_par_breadth_first_search(c: &mut Criterion) {
 	// println!("constructing graph of size = {} Mb", ((SIMPLE_NODE_COUNT * std::mem::size_of::<Node<usize, usize, usize>>()) + (SIMPLE_NODE_COUNT * SIMPLE_NODE_DEGREE * std::mem::size_of::<Edge<usize, usize, usize>>())) / 1000_000);
 	fn digraph_par_bfs() {
 		let t = SIMPLE_GRAPH.node(&rand_range(0, SIMPLE_NODE_COUNT)).unwrap();
-		SIMPLE_GRAPH.par_breadth_first(&rand_range(0, SIMPLE_NODE_COUNT),
-	|e| if *t == e.target() { Finish } else { Include });
+		let closure = | e: &Arc<Edge<usize, usize, Null>> | {
+
+			let mut n = e.target().load();
+
+			if primes::is_prime(n as u64) == true {
+				n = n / 2;
+				e.target().store(n);
+			}
+			if *t == e.target() {
+				Finish
+			} else {
+				Include
+			}
+		};
+		SIMPLE_GRAPH.par_breadth_first(&rand_range(0, SIMPLE_NODE_COUNT), closure);
 	}
 	println!("graph node count = {}", SIMPLE_GRAPH.node_count());
 	println!("graph edge count = {}", SIMPLE_GRAPH.edge_count());
@@ -112,14 +140,11 @@ fn digraph_find_shortest_path(c: &mut Criterion) {
 
 fn digraph_max_flow(c: &mut Criterion) {
 	fn digraph_mf() {
-		maximum_flow_edmonds_karp(&FLOW_GRAPH, rand_range(0, 1000), rand_range(0, 1000));
+		let g = create_graph_flow();
+		maximum_flow_edmonds_karp(&g, rand_range(0, FLOW_NODE_COUNT), rand_range(0, FLOW_NODE_COUNT));
 	}
-	println!("graph node count = {}", FLOW_GRAPH.node_count());
-	println!("graph edge count = {}", FLOW_GRAPH.edge_count());
-	println!("graph average degree = {}", FLOW_GRAPH.edge_count() / FLOW_GRAPH.node_count());
-	println!("sizeof graph = {} Mb", FLOW_GRAPH.bytesize() as f64 / 1000_000.0);
     c.bench_function("maximum flow edmonds karp", |b| b.iter(|| black_box(digraph_mf())));
 }
 
-criterion_group!(benches, digraph_breadth_construction, digraph_breadth_first_search, digraph_par_breadth_first_search, digraph_find_shortest_path, digraph_max_flow);
+criterion_group!(benches, digraph_breadth_first_search, digraph_par_breadth_first_search);
 criterion_main!(benches);
